Feat: create crudoperations for transaction and transfer with each their both endpoint

+++ sql/transfer.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS transfer(
     transfer_id serial PRIMARY KEY,
     account_id_recipient integer NOT NULL,
-    amount decimal(10, 2) NOT NULL,
+    amount Double NOT NULL,
     transfer_reason varchar(50) NOT NULL,
     effective_date TIMESTAMP NOT NULL,
     registration_date TIMESTAMP NOT NULL,
